+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+-   Autocompletion scene.
+
+### Changed
+
+-   Existing autocompletion have been moved from Input.cs to Autocompletion.cs.
+
 ## [1.8.0-beta 2023-11-21]
 
 ### Added